refactor states in TimeRangeSelector and DaysOfService
